requirements.txt
asgiref==3.11.1
azure-ai-textanalytics==5.4.0
azure-cognitiveservices-speech==1.49.1
azure-common==1.1.28
azure-core==1.39.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.3
azure-monitor-opentelemetry==1.8.7
azure-monitor-opentelemetry-exporter==1.0.0b51
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
Flask==3.1.3
gunicorn==25.3.0
idna==3.12
importlib_metadata==8.7.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msal==1.36.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
opentelemetry-api==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-dbapi==0.61b0
opentelemetry-instrumentation-django==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-psycopg2==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-urllib==0.61b0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
packaging==26.1
psutil==7.2.2
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==1.17.3
zipp==3.23.1
